pyproject.toml
[tool.poetry]
name = "algo-deployment"
version = "0.1.0"
description = ""
authors = ["Alteian <Alteian@proton.me>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = "^0.195.2"
starlette = "^0.30.0"
nbformat = "^5.9.1"
nbconvert = "^7.6.0"
uvicorn = "^0.23.0"
pandas = "^2.0.3"
pydantic = "^2.0.3"
ipython = "^8.14.0"
scikit-learn = "^1.3.0"
gunicorn = "^21.2.0"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.278"

[tool.ruff]
ignore-init-module-imports = true
line-length = 79
select = [
  "A",
  "ASYNC100",
  "ASYNC101",
  "ASYNC102",
  "B",
  "BLE",
  "C4",
  "COM",
  "D",
  "D2",
  "D3",
  "D4",
  "DTZ",
  "E",
  "ERA",
  "EXE",
  "F",
  "FLY",
  "G",
  "I",
  "ICN001",
  "INP",
  "INT",
  "ISC",
  "N",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "TD",
  "SIM",
  "SLF",
  "T10",
  "T20",
  "TCH",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
ignore = [
  "D1",
  "D203",
  "D213",
  "PGH003",
  "PLR09",
  "RUF009",
  "SLF001",
  "TD001",
  "TD002",
  "TD003",
  "TRY003",
  "RUF012",
  "D401",
  "EXE002",
  "T201",
  "A002"
]
target-version = "py38"
exclude = [
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".tox",
  ".venv",
  "__pycached__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "*/test_model_load.py"
]

[tool.ruff.per-file-ignores]
"*/converted_script.py" = ["INP001", "UP009", "E402", "COM812", "T201", "E501", "PTH123"]
"__init__.py" = ["F401"]

[tool.ruff.isort]

[tool.black]
line-length = 79
target-version = ['py311']
preview = true
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
